+++ src/core/APIInterface.ts
@@ -3,12 +3,14 @@ import { cities } from "../data/city_data";
 
 class APIInterface {
     private readonly _baseWeatherUrl: string;
+    private readonly _baseOpendStreetMapUrl: string;
     private readonly _queryString: string;
     private readonly _chunkSize: number;
     private readonly _weatherAPIKey: string;
 
     constructor() {
         this._baseWeatherUrl = "https://api.weatherbit.io/v2.0/";
+        this._baseOpendStreetMapUrl = "https://nominatim.openstreetmap.org/search";
         this._weatherAPIKey = import.meta.env.VITE_WEATHER_API_KEY;
         this._chunkSize = 100;
         this._queryString = cities.map((city) => city.city_id).join(",");
@@ -61,7 +63,6 @@ class APIInterface {
     async fetchAlertsForCities(cityChunk) {
         const queryString = cityChunk.map((city) => city.city_id).join(",");
         const url = `${this._baseWeatherUrl}alerts?cities=${queryString}&key=${this._weatherAPIKey}`;
-        console.log("url", url);
         try {
             const response = await fetch(url);
             if (!response.ok) {
@@ -97,6 +98,34 @@ class APIInterface {
         console.log("All", allResponses);
         // Combine all responses into a single result
         return allResponses.flat(); // Flatten the array if needed
+    }
+
+    async fetchCityData(city_name: string, lat: number, lon: number) {
+        let city_data = null;
+        const url = `${this._baseOpendStreetMapUrl}?city=${city_name}&country=PH&type=town&format=json&polygon_geojson=1`;
+        try {
+            const response = await fetch(url);
+            if (!response.ok) {
+                throw new Error(`Error fetching data: ${response.statusText}`);
+            }
+            city_data = await response.json();
+            if (city_data?.length > 0) {
+                let i = 0;
+                for (i = 0; i < city_data?.length; i++) {
+                    const item = city_data[i];
+                    if (
+                        item?.lat?.split(".")[0] === lat.toString().split(".")[0] &&
+                        item?.lon?.split(".")[0] === lon.toString().split(".")[0] &&
+                        (item?.type === "town" || item?.type === "administrative")
+                    ) {
+                        return item;
+                    }
+                }
+            }
+        } catch (error) {
+            console.error("Error fetching weather data:", error);
+            return null; // Return null or handle the error as needed
+        }
     }
 }