+++ src/core/Area.ts
@@ -1,6 +1,19 @@
-import { Billboard, BillboardCollection, Cartesian2, Cartesian3, Event } from "cesium";
+import {
+    Billboard,
+    BillboardCollection,
+    Cartesian2,
+    Cartesian3,
+    Color,
+    DistanceDisplayCondition,
+    Entity,
+    EntityCollection,
+    Event,
+    Math as CesiumMath,
+    PolygonHierarchy
+} from "cesium";
 import { Point } from "geojson";
-import { WeatherType, CapturedCameraProps } from "./common";
+import { AlertType, WeatherType, CapturedCameraProps } from "./common";
+import markerAlert from "../assets/images/flood_alert.png";
 
 interface WeatherConstructorOptions {
     point: Point;
@@ -14,13 +27,27 @@ export class Area {
 
     private _latitude: number;
 
+    private _minLongitude: number;
+
+    private _minLatitude: number;
+
+    private _maxLongitude: number;
+
+    private _maxLatitude: number;
+
     private _cameraProps: CapturedCameraProps;
 
     private _options: WeatherConstructorOptions;
 
     private _type: WeatherType;
 
+    private _alertType: AlertType;
+
+    private _alert: any;
+
     private readonly _billboard: Billboard;
+
+    private cityRange: Entity | undefined;
 
     private _removeEvent: Event.RemoveCallback | undefined;
 
@@ -35,6 +62,8 @@ export class Area {
 
         this._longitude = longitude;
         this._latitude = latitude;
+        this._minLatitude = this._maxLatitude = latitude;
+        this._minLongitude = this._maxLongitude = longitude;
 
         this._cameraProps = null;
 
@@ -45,6 +74,8 @@ export class Area {
         } else {
             this._type = WeatherType.STATE_900;
         }
+
+        this._alertType = AlertType.ALERT_000;
 
         const icon = options.properties?.weather?.icon;
 
@@ -70,7 +101,6 @@ export class Area {
             height: 50,
             scale: 0.8
         });
-
         this._properties = options.properties;
     }
 
@@ -86,6 +116,38 @@ export class Area {
         return this._latitude;
     }
 
+    get minLongitude() {
+        return this._minLongitude;
+    }
+
+    set minLongitude(val: number) {
+        this._minLongitude = val;
+    }
+
+    get maxLongitude() {
+        return this._maxLongitude;
+    }
+
+    set maxLongitude(val: number) {
+        this._maxLongitude = val;
+    }
+
+    get minLatitude() {
+        return this._minLatitude;
+    }
+
+    set minLatitude(val: number) {
+        this._minLatitude = val;
+    }
+
+    get maxLatitude() {
+        return this._maxLatitude;
+    }
+
+    set maxLatitude(val: number) {
+        this._maxLatitude = val;
+    }
+
     get billboard() {
         return this._billboard;
     }
@@ -96,6 +158,18 @@ export class Area {
 
     set cameraProps(props: CapturedCameraProps) {
         this._cameraProps = props;
+    }
+
+    get alertType() {
+        return this._alertType;
+    }
+
+    get alert() {
+        return this._alert;
+    }
+
+    set alert(val: any) {
+        this._alert = val;
     }
 
     highlight() {
@@ -122,13 +196,61 @@ export class Area {
 
     showHide(show: boolean) {
         this._billboard.show = show;
+        if (this.cityRange) {
+            this.cityRange.show = show;
+        }
     }
 
     getShow() {
         return this._billboard.show;
     }
 
-    geWeatherStatus() {
+    getWeatherStatus() {
         return this._properties?.weather?.description;
+    }
+
+    setAlert(alert: any) {
+        this._alertType = AlertType.ALERT_1000;
+        this._alert = alert;
+        this._billboard.image = markerAlert;
+    }
+
+    drawCityRange() {
+        const radius = this.getRadius(this._minLongitude, this._minLatitude, this._maxLongitude, this._maxLatitude, 0);
+        const circlePositions = Cartesian3.fromDegreesArray(
+            this.generateCirclePoints(this._longitude, this._latitude, radius)
+        );
+        const geoTech = window.geoTech;
+        this.cityRange = geoTech.viewer.entities.add({
+            polygon: {
+                hierarchy: new PolygonHierarchy(circlePositions),
+                material: Color.RED.withAlpha(0.1), // Semi-transparent red
+                outline: true,
+                outlineColor: Color.RED,
+                distanceDisplayCondition: new DistanceDisplayCondition(0, 1e6)
+            }
+        });
+    }
+
+    generateCirclePoints(centerLon, centerLat, radius) {
+        const positions = [];
+        const numPoints = 360; // Number of points to approximate the circle
+        for (let i = 0; i < numPoints; i++) {
+            const angle = CesiumMath.toRadians(i); // Convert degrees to radians
+            const lon = centerLon + (radius / 111319) * Math.cos(angle); // Adjust longitude
+            const lat = centerLat + (radius / 111319) * Math.sin(angle); // Adjust latitude
+            positions.push(lon, lat);
+        }
+        return positions;
+    }
+
+    getRadius(minX, minY, maxX, maxY, height = 0) {
+        const minCartesian = Cartesian3.fromDegrees(minX, minY, height);
+        const maxCartesian = Cartesian3.fromDegrees(maxX, maxY, height);
+
+        // Calculate the center and radius
+        const centerCartesian = Cartesian3.midpoint(minCartesian, maxCartesian, new Cartesian3());
+        const radius = Cartesian3.distance(centerCartesian, maxCartesian);
+        return radius;
     }
 }