+++ src/index.ts
@@ -5,7 +5,7 @@ import render from "./render";
 import { setAssetPath } from "@esri/calcite-components/dist/components";
 
 import "@esri/calcite-components/dist/calcite/calcite.css";
-setAssetPath("https://js.arcgis.com/calcite-components/1.2.0/assets");
+setAssetPath("https://js.arcgis.com/calcite-components/2.13.2/assets");
 
 const geoTech = new GeoTech();
 window.geoTech = geoTech;